Refactored the architecture and optimized the code implementation
