+++ README.md
@@ -66,13 +66,6 @@ Add the JitPack repository to your build file
     }
     ```
 
-### Latest supported properties configuration
-
-``` properties
-kotlin.version=2.0.21
-compose.version=1.7.3
-```
-
 ### Usage/Examples
 
 #### Resources file
